ノード状態の初期値を-1or1に

+++ reservoir_ESN/reservoir_layer.h
@@ -38,7 +38,7 @@ public:
 	//double pa;
 
 	std::vector<double> a;
-	std::vector<int> b;
+	std::vector<double> b;
